Fix problem with reactive helper
Since Meteor's reactive helpers are invoked
multiple times, each time the page is queried,
the current state needs to be chacked again.
This leads to ugly clipping, which is now covered
with a spinner. The spinner transition time can
be set up in the settings. A value of 0 deactivates
the spinner.